Update app configuration
Created three environments for the app to run in. Development,
Production, and Testing. Before this refactoring the app always ran in a
production environment which is OK for a while, but can cause issues
when developing. For instance: The DB would be wiped out everytime the
test suite was ran which could cause dev frustration and the need to
recreate the DB a lot. Furthermore, libraries may contain environment
specific configurations depending on which environment the app is
running in, and it is best to take advantage of that.

Removed the method and associated test that checked to see if the DB was
seeded. This was helpful when there was a single environment for the app
to run in. However, now I don't see it being necessary as the dev
environment can be seeded once and will not be wiped anymore when tests
are ran. Plus removing non-app specific code improves the code base. The
friendly dev message it provided is nice, however it would get buried in
the stack trace. I think it would be more helpful to let the SQLalchemy
db exception propagate to the top which makes it pretty clear that the db
is the issue.

Finally, secured all env vars. There are now two env files to store
environment variables in for developing. The first is a public file that
can be used to store, well, public env vars `.flaskenv`. And the second
is a private file that is used to store private (secret) env vars.

Currently there are only two env vars needed for this app to run. The
first is flask specific and is stored in `.flaskenv` which tells flask
what environment to use when developing. For example: when using any
Flask commands such as `flask run`. The second is the secret_key which
is stored in the private `.env` file.

Updated the README to include add instructions for copying the
`.env.example` file.

Fixes codeforpdx/dwellingly-app/issues/252
Fixes codeforpdx/dwellingly-app/issues/307